backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]
websockets
pydantic
requests
validators
tldextract
python-whois
httpx
sqlalchemy>=2.0.0
pymysql>=1.1.0
feedparser>=6.0.0
python-dotenv
apscheduler>=3.10.0
cryptography>=41.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0
pandas>=2.0.0
numpy==1.26.4
joblib>=1.3.0
